Moved hGraphQL out of payments
