+++ Projects/Payment/Sources/Models/PaymentData.swift
@@ -2,27 +2,58 @@ import Contracts
 import Foundation
 import hCore
 import hCoreUI
-import hGraphQL
 
 public struct PaymentData: Codable, Equatable, Hashable, Sendable {
     let id: String
-    let payment: PaymentStack
+    public let payment: PaymentStack
     let status: PaymentStatus
     let contracts: [ContractPaymentDetails]
-    let discounts: [Discount]
+    public let discounts: [Discount]
     let paymentDetails: PaymentDetails?
     //had to add as an array since we can't nest same struct type here
     let addedToThePayment: [PaymentData]?
 
-    struct PaymentStack: Codable, Equatable, Hashable, Sendable {
+    public init(
+        id: String,
+        payment: PaymentStack,
+        status: PaymentStatus,
+        contracts: [ContractPaymentDetails],
+        discounts: [Discount],
+        paymentDetails: PaymentDetails?,
+        addedToThePayment: [PaymentData]?
+    ) {
+        self.id = id
+        self.payment = payment
+        self.status = status
+        self.contracts = contracts
+        self.discounts = discounts
+        self.paymentDetails = paymentDetails
+        self.addedToThePayment = addedToThePayment
+    }
+
+    public struct PaymentStack: Codable, Equatable, Hashable, Sendable {
         let gross: MonetaryAmount
         let net: MonetaryAmount
         let carriedAdjustment: MonetaryAmount?
         let settlementAdjustment: MonetaryAmount?
-        let date: ServerBasedDate
+        public let date: ServerBasedDate
+
+        public init(
+            gross: MonetaryAmount,
+            net: MonetaryAmount,
+            carriedAdjustment: MonetaryAmount?,
+            settlementAdjustment: MonetaryAmount?,
+            date: ServerBasedDate
+        ) {
+            self.gross = gross
+            self.net = net
+            self.carriedAdjustment = carriedAdjustment
+            self.settlementAdjustment = settlementAdjustment
+            self.date = date
+        }
     }
 
-    enum PaymentStatus: Codable, Equatable, Hashable, Sendable {
+    public enum PaymentStatus: Codable, Equatable, Hashable, Sendable {
         case upcoming
         case success
         case pending
@@ -47,21 +78,45 @@ public struct PaymentData: Codable, Equatable, Hashable, Sendable {
         }
     }
 
-    struct ContractPaymentDetails: Codable, Equatable, Identifiable, Hashable, Sendable {
-        let id: String
+    public struct ContractPaymentDetails: Codable, Equatable, Identifiable, Hashable, Sendable {
+        public let id: String
         let title: String
         let subtitle: String?
         let amount: MonetaryAmount
         let periods: [PeriodInfo]
+
+        public init(id: String, title: String, subtitle: String?, amount: MonetaryAmount, periods: [PeriodInfo]) {
+            self.id = id
+            self.title = title
+            self.subtitle = subtitle
+            self.amount = amount
+            self.periods = periods
+        }
     }
 
-    struct PeriodInfo: Codable, Equatable, Identifiable, Hashable, Sendable {
-        let id: String
+    public struct PeriodInfo: Codable, Equatable, Identifiable, Hashable, Sendable {
+        public let id: String
         let from: ServerBasedDate
         let to: ServerBasedDate
         let amount: MonetaryAmount
         let isOutstanding: Bool
         let desciption: String?
+
+        public init(
+            id: String,
+            from: ServerBasedDate,
+            to: ServerBasedDate,
+            amount: MonetaryAmount,
+            isOutstanding: Bool,
+            desciption: String?
+        ) {
+            self.id = id
+            self.from = from
+            self.to = to
+            self.amount = amount
+            self.isOutstanding = isOutstanding
+            self.desciption = desciption
+        }
 
         @MainActor
         var fromToDate: String {
@@ -69,13 +124,13 @@ public struct PaymentData: Codable, Equatable, Hashable, Sendable {
         }
     }
 
-    struct PaymentDetails: Codable, Equatable, Hashable, Sendable {
+    public struct PaymentDetails: Codable, Equatable, Hashable, Sendable {
         typealias KeyValue = (key: String, value: String)
         let paymentMethod: String
         let account: String
         let bank: String
 
-        init(paymentMethod: String, account: String, bank: String) {
+        public init(paymentMethod: String, account: String, bank: String) {
             self.paymentMethod = paymentMethod
             self.account = account
             self.bank = bank
@@ -93,14 +148,34 @@ public struct PaymentData: Codable, Equatable, Hashable, Sendable {
 }
 
 public struct Discount: Codable, Equatable, Identifiable, Hashable, Sendable {
-    public let id = UUID().uuidString
-    let code: String
-    let amount: MonetaryAmount?
+    public let id: String
+    public let code: String
+    public let amount: MonetaryAmount?
     let title: String?
     let listOfAffectedInsurances: [AffectedInsurance]
     let validUntil: ServerBasedDate?
     let canBeDeleted: Bool
     let discountId: String
+
+    public init(
+        id: String,
+        code: String,
+        amount: MonetaryAmount?,
+        title: String?,
+        listOfAffectedInsurances: [AffectedInsurance],
+        validUntil: ServerBasedDate?,
+        canBeDeleted: Bool,
+        discountId: String
+    ) {
+        self.id = id
+        self.code = code
+        self.amount = amount
+        self.title = title
+        self.listOfAffectedInsurances = listOfAffectedInsurances
+        self.validUntil = validUntil
+        self.canBeDeleted = canBeDeleted
+        self.discountId = discountId
+    }
 
     @MainActor
     var isValid: Bool {
@@ -120,6 +195,11 @@ public struct Discount: Codable, Equatable, Identifiable, Hashable, Sendable {
 public struct AffectedInsurance: Codable, Equatable, Identifiable, Hashable, Sendable {
     public let id: String
     let displayName: String
+
+    public init(id: String, displayName: String) {
+        self.id = id
+        self.displayName = displayName
+    }
 }
 
 extension PaymentData: TrackingViewNameProtocol {
