+++ Projects/Payment/Sources/PaymentStore.swift
@@ -3,11 +3,10 @@ import Contracts
 import Foundation
 import PresentableStore
 import hCore
-import hGraphQL
 
 public struct PaymentState: StateProtocol {
-    var paymentData: PaymentData?
-    var ongoingPaymentData: [PaymentData] = []
+    public var paymentData: PaymentData?
+    public var ongoingPaymentData: [PaymentData] = []
     var paymentDiscountsData: PaymentDiscountsData?
     public var paymentStatusData: PaymentStatusData? = nil
     var paymentHistory: [PaymentHistoryListData] = []
