+++ Projects/Payment/Sources/Screens/PaymentDetails/PaymentDetailsView.swift
@@ -275,6 +275,7 @@ struct PaymentDetails_Previews: PreviewProvider {
             ],
             discounts: [
                 .init(
+                    id: "CODE",
                     code: "CODE",
                     amount: .sek(100),
                     title: "Title",
@@ -286,6 +287,7 @@ struct PaymentDetails_Previews: PreviewProvider {
                     discountId: "CODE"
                 ),
                 .init(
+                    id: "CODE2",
                     code: "CODE2",
                     amount: .sek(99),
                     title: "Title1",
@@ -297,6 +299,7 @@ struct PaymentDetails_Previews: PreviewProvider {
                     discountId: "CODE2"
                 ),
                 .init(
+                    id: "FRIENDS",
                     code: "MY CODE",
                     amount: .sek(30),
                     title: "3 friends invited",
