+++ Projects/Payment/Sources/Screens/PaymentsView.swift
@@ -3,7 +3,6 @@ import PresentableStore
 import SwiftUI
 import hCore
 import hCoreUI
-import hGraphQL
 
 public struct PaymentsView: View {
     @PresentableStore var store: PaymentStore
