+++ Projects/Payment/Sources/Service/CampaignService/CampaignClient.swift
@@ -1,6 +1,5 @@
 import Foundation
 import hCore
-import hGraphQL
 
 @MainActor
 public protocol hCampaignClient {
@@ -8,13 +7,13 @@ public protocol hCampaignClient {
     func add(code: String) async throws
 }
 
-enum CampaignError: Error {
+public enum CampaignError: Error {
     case userError(message: String)
     case notImplemented
 }
 
 extension CampaignError: LocalizedError {
-    var errorDescription: String? {
+    public var errorDescription: String? {
         switch self {
         case let .userError(message):
             return message
