+++ Projects/Payment/Sources/Service/PaymentService/Protocols/PaymentClient.swift
@@ -9,12 +9,12 @@ public protocol hPaymentClient {
     func getConnectPaymentUrl() async throws -> URL
 }
 
-enum PaymentError: Error {
+public enum PaymentError: Error {
     case missingDataError(message: String)
 }
 
 extension PaymentError: LocalizedError {
-    var errorDescription: String? {
+    public var errorDescription: String? {
         switch self {
         case let .missingDataError(message): return message
         }
