+++ Projects/Payment/Tests/StoreTests/StoreDiscountsTests.swift
@@ -21,6 +21,7 @@ final class StoreDiscountsTests: XCTestCase {
         let discountsData: PaymentDiscountsData = .init(
             discounts: [
                 .init(
+                    id: "id",
                     code: "code",
                     amount: .init(amount: "30", currency: "SEK"),
                     title: "title",
